Refactor out some of the shared API sidebar code.
