+++ src/navigation/GraphQLCategorySidebar.tsx
@@ -1,69 +1,15 @@
-import React, { useMemo, useRef } from "react";
+import React from "react";
 import { useSiteData } from "react-static";
-import styled from "styled-components";
 
 import {
   StyledApiLink,
   MethodDiv,
   LabelsDiv,
   StyledLabelSpan,
-} from "./ApiSidebar";
+} from "./api/styles";
 import { Category } from "src/graphql/category";
 import { CategoryEntry, ProcessedSchema } from "src/graphql/types";
-
-interface CollapsibleApiMenuProps {
-  name: string;
-}
-
-const CollapsibleApiMenuHeading = styled.div`
-  display: block;
-  height: fit-content;
-  align-items: center;
-  width: auto !important;
-  font-size: 14px;
-  line-height: 21px;
-  position: relative !important;
-  padding: 8px 20% 8px 12px !important;
-  cursor: pointer;
-  &:hover {
-    background-color: #e7edee;
-  }
-`;
-
-let collapsibleMenuIdx = 0;
-
-// TODO: Factor out
-const CollapsibleApiMenu: React.FC<CollapsibleApiMenuProps> = (props) => {
-  const parent = useRef(null);
-  const id = useMemo(() => `#mm-collapsible-menu-${collapsibleMenuIdx++}`, []);
-
-  function open(e: React.MouseEvent) {
-    e.preventDefault();
-    jQuery(parent.current).toggleClass("mm-opened");
-  }
-
-  return (
-    <li className="mm-vertical" ref={parent}>
-      <CollapsibleApiMenuHeading
-        className="mm-next mm-fullsubopen"
-        data-target={id}
-        onClick={open}
-      >
-        {props.name}
-      </CollapsibleApiMenuHeading>
-      <div
-        className="mm-panel mm-vertical"
-        style={{
-          marginLeft: "12px",
-          borderLeft: "1px solid #003b45",
-        }}
-        id={id}
-      >
-        <ul className="nav-onpage mm-listview mm-vertical">{props.children}</ul>
-      </div>
-    </li>
-  );
-};
+import CollapsibleApiMenu from "./api/CollapsibleApiMenuItem";
 
 export default () => {
   const { graphql } = useSiteData<{
@@ -106,7 +52,7 @@ export default () => {
         <li key={item.url}>
           <StyledApiLink
             to={item.url}
-            clicked={window.location.pathname === item.url}
+            $clicked={window.location.pathname === item.url}
           >
             <MethodDiv>
               {item.name}
@@ -128,7 +74,7 @@ export default () => {
             <li key={item.url}>
               <StyledApiLink
                 to={item.url}
-                clicked={window.location.pathname === item.url}
+                $clicked={window.location.pathname === item.url}
               >
                 {item.name}
               </StyledApiLink>
