+++ src/navigation/GraphQLReferenceSidebar.tsx
@@ -1,63 +1,9 @@
-import React, { useMemo, useRef } from "react";
+import React from "react";
 import { useSiteData } from "react-static";
-import styled from "styled-components";
 
-import { StyledApiLink } from "./ApiSidebar";
+import { StyledApiLink } from "./api/styles";
 import { ProcessedSchema, BaseDefinition } from "src/graphql/types";
-
-interface CollapsibleApiMenuProps {
-  name: string;
-}
-
-const CollapsibleApiMenuHeading = styled.div`
-  display: block;
-  height: fit-content;
-  align-items: center;
-  width: auto !important;
-  font-size: 14px;
-  line-height: 21px;
-  position: relative !important;
-  padding: 8px 20% 8px 12px !important;
-  cursor: pointer;
-  &:hover {
-    background-color: #e7edee;
-  }
-`;
-
-let collapsibleMenuIdx = 1000;
-
-// TODO: Factor out
-const CollapsibleApiMenu: React.FC<CollapsibleApiMenuProps> = (props) => {
-  const parent = useRef(null);
-  const id = useMemo(() => `#mm-collapsible-menu-${collapsibleMenuIdx++}`, []);
-
-  function open(e: React.MouseEvent) {
-    e.preventDefault();
-    jQuery(parent.current).toggleClass("mm-opened");
-  }
-
-  return (
-    <li className="mm-vertical" ref={parent}>
-      <CollapsibleApiMenuHeading
-        className="mm-next mm-fullsubopen"
-        data-target={id}
-        onClick={open}
-      >
-        {props.name}
-      </CollapsibleApiMenuHeading>
-      <div
-        className="mm-panel mm-vertical"
-        style={{
-          marginLeft: "12px",
-          borderLeft: "1px solid #003b45",
-        }}
-        id={id}
-      >
-        <ul className="nav-onpage mm-listview mm-vertical">{props.children}</ul>
-      </div>
-    </li>
-  );
-};
+import CollapsibleApiMenu from "./api/CollapsibleApiMenuItem";
 
 export default () => {
   const { graphql } = useSiteData<{
@@ -70,12 +16,10 @@ export default () => {
   const mutations = Object.values(graphql.mutations).sort((a, b) =>
     a.name.localeCompare(b.name)
   );
-  const objects = Object.values(graphql.objects).sort((a, b) =>
-    a.name.localeCompare(b.name)
-  );
-  const inputObjects = Object.values(graphql.inputObjects).sort((a, b) =>
-    a.name.localeCompare(b.name)
-  );
+  const objects = [
+    ...Object.values(graphql.objects),
+    ...Object.values(graphql.inputObjects),
+  ].sort((a, b) => a.name.localeCompare(b.name));
   const enums = Object.values(graphql.enums).sort((a, b) =>
     a.name.localeCompare(b.name)
   );
@@ -96,7 +40,7 @@ export default () => {
           <li key={item.url}>
             <StyledApiLink
               to={item.url}
-              clicked={window.location.pathname === item.url}
+              $clicked={window.location.pathname === item.url}
             >
               {item.name}
             </StyledApiLink>
